+++ library/core/src/main/java/com/google/android/exoplayer2/SimpleExoPlayer.java
@@ -2098,6 +2098,11 @@ public class SimpleExoPlayer extends BasePlayer
       analyticsCollector.onVideoFrameProcessingOffset(totalProcessingOffsetUs, frameCount);
     }
 
+    @Override
+    public void onVideoCodecError(Exception videoCodecError) {
+      analyticsCollector.onVideoCodecError(videoCodecError);
+    }
+
     // AudioRendererEventListener implementation
 
     @Override
@@ -2154,6 +2159,11 @@ public class SimpleExoPlayer extends BasePlayer
     @Override
     public void onAudioSinkError(Exception audioSinkError) {
       analyticsCollector.onAudioSinkError(audioSinkError);
+    }
+
+    @Override
+    public void onAudioCodecError(Exception audioCodecError) {
+      analyticsCollector.onAudioCodecError(audioCodecError);
     }
 
     // TextOutput implementation